logging/viz: Move `logs_to_dot` and utils into Dagger
